pwgo_helper_package/requirements.txt
python-dotenv
click
pymysql
aiomysql
mysql-replication
path
requests
beautifulsoup4
py-linq
aiobotocore
fs
pillow
pyexiv2==2.5.0
pyicloud_ipd==0.10.1
slack_sdk
pid
